roles permissions step-2

+++ app/Http/Controllers/Admin/TasksController.php
@@ -5,33 +5,33 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
-use App\Models\Task;
+use Symfony\Component\HttpFoundation\Response;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\Admin\StoreTaskRequest;
 use App\Http\Requests\Admin\UpdateTaskRequest;
+use App\Models\Task;
 
 class TasksController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+
     public function index()
     {
-        $tasks = Task::paginate(15);
+        abort_if(Gate::denies('task_management_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        $tasks = Task::paginate();
 
         return view('admin.tasks.index',compact('tasks'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
+
     public function create()
     {
+        abort_if(Gate::denies('task_management_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.tasks.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
+
     public function store(StoreTaskRequest $request)
     {
         Task::create($request->validated());
@@ -39,25 +39,23 @@ class TasksController extends Controller
         return redirect()->route('admin.tasks.index');
     }
 
-    /**
-     * Display the specified resource.
-     */
+
     public function show(Task $task)
     {
+        abort_if(Gate::denies('task_management_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.tasks.show',compact('task'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
+
     public function edit(Task $task)
     {
+        abort_if(Gate::denies('task_management_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.tasks.edit',compact('task'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
+
     public function update(UpdateTaskRequest $request, Task $task)
     {
         $task->update($request->validated());
@@ -65,11 +63,11 @@ class TasksController extends Controller
         return redirect()->route('admin.tasks.show',$task->id);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+
     public function destroy(Task $task)
     {
+        abort_if(Gate::denies('task_management_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $task->delete();
 
         return redirect()->route('admin.tasks.index');
